refactor: 테스트 공통 함수 분리
